requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
joblib==1.4.2
vaderSentiment==3.3.2
python-dateutil==2.9.0.post0
